refactor(depgraph): split snapshot.rs into focused submodules
Convert crates/zccache-depgraph/src/snapshot.rs (2068 lines) into a
directory module so every file stays under 1,000 LOC. Production code
is split into a tiny mod.rs (types + DepGraph conversion methods) and
persistence.rs (file I/O), and the test module is split per concern.
No behavioral changes — purely a code split. Public API is unchanged:
every previously-pub item at zccache_depgraph::snapshot::* is preserved
via pub use re-exports.

Layout (new files):
- snapshot/mod.rs: public snapshot types (DepGraphSnapshot,
  FileEntrySnapshot, IncludeDirectiveSnapshot, ContextEntrySnapshot,
  SnapshotStats), SnapshotError, magic + version constants,
  impl DepGraph conversion methods (to_snapshot / from_snapshot),
  paths_to_strings / strings_to_paths helpers.
- snapshot/persistence.rs: save_to_file, load_from_file, classify_load,
  DepGraphLoadOutcome, depgraph_file_path.
- snapshot/tests/mod.rs: declares submodules + shared helpers
  (test_path, make_ctx, dummy_hash, always_fresh).
- snapshot/tests/round_trip.rs: per-field snapshot <-> load roundtrips
  (empty, populated, last_file_hashes, artifact_key Some/None, unicode
  defines, has_computed_includes, IncludeKind::Computed, empty strings,
  exact byte equality, stats reset).
- snapshot/tests/persistence.rs: file I/O + on-disk format edge cases
  (bad magic, version mismatch, truncated payload, missing file,
  atomic tmp cleanup, overwrite, zero-length payload, header too short,
  trailing garbage, payload-length overflow, classify_load matrix).
- snapshot/tests/behavioral.rs: cross-cutting behavior (cache-hit
  recovery, context-key consistency, unicode paths, double-roundtrip
  idempotency, overlapping contexts, all ContextState variants, bit-flip
  detection, register-after-load, GC on save, concurrent save/load,
  large-graph stress).
- snapshot/README.md, snapshot/tests/README.md: hook-required.

LOC after split:
- mod.rs: 287
- persistence.rs: 212
- tests/mod.rs: 45
- tests/round_trip.rs: 525
- tests/persistence.rs: 328
- tests/behavioral.rs: 760

Tests: zccache-depgraph --lib still passes 341 tests (40 snapshot::*
tests, identical to before the split). Clippy clean on the workspace.
